Fix initialization of volume on startup
The extra event handler that claimed to fix something was actually
causing problems in Chrome on my machine: it would reset the volume
to 50% every other time the game was reloaded.

The SugarCube documentation states that calling `Settings.load()`
and `Settings.save()` is not necessary unless you're doing unusual
things, so I removed that as well.