Add Cherry and Cloud's special abilities as text
For Cherry's Chaotic Luck, the main thing the player should know
is that it kicks in every time you reach a new layer. The exact
randomization doesn't have to be explained since the player isn't
executing it, unlike the CYOA.

For Cloud's connections, I transcribed the text and added it in
his post-hire notification. Showing it before he's hired would
complicate the layout. Maybe it's better to show it at the moment
the player considers whether to take out a loan or not, but that
requires a lot more coding, so I opted to not do that right now.