Move item descriptions from variables into passages
Not all descriptions are in passages yet: the duplicate items from
the layer 7 vending machine are defined using constants containing
descriptions. These should be given unique names before we can move
the description.

Giving them a unique name is also needed to make the L7 versions
actually lighter than the surface versions. Currently this is not
possible because there is only one kind of rations and only one
kind of medkit in the inventory.